+++ CHANGES.rst
@@ -9,6 +9,7 @@
 * update defaults for filter_outliers
 * use vsearch 2.0.3
 * fix error in fill_lonely raised by node with no parent
+* added `deenurp orientate_sequences --out_notmatched_taxids`
 
 0.1.8
 ======